requirements.txt
Package                   Version
------------------------- --------------------
absl-py                   2.0.0
antlr4-python3-runtime    4.9.3
anyio                     4.0.0
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 2.4.1
async-lru                 2.0.4
attrs                     23.1.0
Babel                     2.13.1
beautifulsoup4            4.12.2
bleach                    6.1.0
build                     1.0.3
cachetools                5.3.2
certifi                   2023.7.22
cffi                      1.16.0
charset-normalizer        3.3.2
colorama                  0.4.6
comm                      0.2.0
contourpy                 1.2.0
cycler                    0.12.1
debugpy                   1.8.0
decorator                 5.1.1
defusedxml                0.7.1
deprecation               2.1.0
docutils                  0.20.1
exceptiongroup            1.1.3
executing                 2.0.1
fastjsonschema            2.18.1
filelock                  3.13.1
fonttools                 4.44.1
fqdn                      1.5.1
fsspec                    2023.10.0
gitdb                     4.0.11
GitPython                 3.1.40
google-auth               2.23.4
google-auth-oauthlib      1.1.0
grpcio                    1.59.2
hydra-core                1.3.2
idna                      3.4
importlib-metadata        6.8.0
ipykernel                 6.26.0
ipython                   8.17.2
ipywidgets                8.1.1
isoduration               20.11.0
jaraco.classes            3.3.0
jedi                      0.19.1
Jinja2                    3.1.2
json5                     0.9.14
jsonpointer               2.4
jsonschema                4.19.2
jsonschema-specifications 2023.7.1
jupyter                   1.0.0
jupyter_client            8.6.0
jupyter-console           6.6.3
jupyter_core              5.5.0
jupyter-events            0.9.0
jupyter-lsp               2.2.0
jupyter_server            2.10.0
jupyter_server_terminals  0.4.4
jupyterlab                4.0.8
jupyterlab-pygments       0.2.2
jupyterlab_server         2.25.1
jupyterlab-widgets        3.0.9
keyring                   24.3.0
kiwisolver                1.4.5
Markdown                  3.5.1
markdown-it-py            3.0.0
MarkupSafe                2.1.3
matplotlib                3.8.1
matplotlib-inline         0.1.6
mdurl                     0.1.2
mistune                   3.0.2
more-itertools            10.1.0
mpmath                    1.3.0
nbclient                  0.9.0
nbconvert                 7.11.0
nbformat                  5.9.2
nest-asyncio              1.5.8
networkx                  3.2.1
nh3                       0.2.14
notebook                  7.0.6
notebook_shim             0.2.3
numpy                     1.22.4
oauthlib                  3.2.2
omegaconf                 2.3.0
opencv-python             4.8.1.78
overrides                 7.4.0
packaging                 23.2
pandas                    1.5.3
pandocfilters             1.5.0
parso                     0.8.3
Pillow                    10.1.0
pip                       22.2.2
pkginfo                   1.9.6
platformdirs              4.0.0
prometheus-client         0.18.0
prompt-toolkit            3.0.41
protobuf                  4.23.4
psutil                    5.9.6
pure-eval                 0.2.2
py-cpuinfo                9.0.0
py4j                      0.10.9.5
pyasn1                    0.5.0
pyasn1-modules            0.3.0
pycparser                 2.21
Pygments                  2.16.1
pyparsing                 3.1.1
pyproj                    3.4.1
pyproject_hooks           1.0.0
pyrasterframes            0.11.1
pyspark                   3.3.2
python-dateutil           2.8.2
python-json-logger        2.0.7
pytz                      2023.3.post1
pywin32                   306
pywin32-ctypes            0.2.2
pywinpty                  2.0.12
PyYAML                    6.0.1
pyzmq                     25.1.1
qtconsole                 5.5.0
QtPy                      2.4.1
readme-renderer           42.0
referencing               0.30.2
requests                  2.31.0
requests-oauthlib         1.3.1
requests-toolbelt         1.0.0
rfc3339-validator         0.1.4
rfc3986                   2.0.0
rfc3986-validator         0.1.1
rich                      13.6.0
rpds-py                   0.12.0
rsa                       4.9
scipy                     1.11.3
seaborn                   0.13.0
Send2Trash                1.8.2
sentry-sdk                1.35.0
setuptools                63.2.0
shapely                   2.0.2
six                       1.16.0
smmap                     5.0.1
sniffio                   1.3.0
soupsieve                 2.5
stack-data                0.6.3
sympy                     1.12
tensorboard               2.15.1
tensorboard-data-server   0.7.2
terminado                 0.18.0
thop                      0.1.1.post2209072238
tinycss2                  1.2.1
tomli                     2.0.1
torch                     2.1.0
torchvision               0.16.0
tornado                   6.3.3
tqdm                      4.66.1
traitlets                 5.13.0
twine                     4.0.2
types-python-dateutil     2.8.19.14
typing_extensions         4.8.0
tzdata                    2023.3
ultralytics               8.0.10
ultralytics-yolo          0.0.1
uri-template              1.3.0
urllib3                   2.1.0
wcwidth                   0.2.10
webcolors                 1.13
webencodings              0.5.1
websocket-client          1.6.4
Werkzeug                  3.0.1
widgetsnbextension        4.0.9
zipp                      3.17.0
